Revert "Simple MEM1 allocator."
MEM1 handling is harder than I thought. While we might be able to place a few small buffes in MEM1 having graphic buffers and SWKBD won't work. Memory management is so complicated that it would be way more simple to rewrie the app from scratch than to implement it now.

This reverts commit 89ad6a49f8648aa00a56b9b5fa6131d879fb3db5.